+++ CHANGELOG.md
@@ -1,4 +1,9 @@
 # Changelog
+
+## **WORK IN PROGRESS**
+
+* (steinwedel) Simplify public holiday detection to a single configurable boolean state, with an object-tree picker in the admin UI
+
 ## 0.0.2 (2026-08-13)
 
 * (steinwedel) initial development version